\name{ALGchangeLog}
\alias{ALGchangeLog}

\title{affylmGUI Change Log}
\description{Write as text the most recent changes from the affylmGUI package changelog.}

\usage{
ALGchangeLog(n=20)
}

\arguments{
  \item{n}{integer, number of lines to write of changelog.}
}

\value{No value is produced, but a number of lines of text are written to standard output.}

\author{Gordon Smyth}

\keyword{documentation}
