\title{Graphical User Interface for the limma microarray package}
\name{affylmGUI}
\alias{affylmGUI}
\alias{AboutaffylmGUI}
\alias{AboutCustomMenu}
\alias{AboutNormalization}
\alias{addMenuItem}
\alias{affylmGUI}
\alias{affyPlotMA}
\alias{affyPlotMAcontrast}
\alias{ChooseCDF}
\alias{ChooseContrastParameterization}
\alias{chooseDir}
\alias{ChooseEbayesStatistic}
\alias{ComputeContrasts}
\alias{ComputeLinearModelFit}
\alias{CopyGraph}
\alias{DeleteContrastParameterization}
\alias{DeleteCustomMenuItem}
\alias{deleteItemFromList}
\alias{DensityPlot}
\alias{DensityPlotAll}
\alias{RNADigestionPlotAll}
\alias{NUSEPlotAll}
\alias{RLEPlotAll}
\alias{evalRcode}
\alias{ExportTopTable}
\alias{ExportHTMLreport}
\alias{ExportNormalizedExpressionValues}
\alias{generalPlotFunction}
\alias{GetComponentsToExportInHTMLreport}
\alias{GetContrast}
\alias{GetContrastNamesForHeatDiagram}
\alias{GetContrastParameterizationName}
\alias{GetContrasts}
\alias{GetCustomMenuItemNum}
\alias{GetDEcutoff}
\alias{GetGeneLabelsOptions}
\alias{GetJpegOrPngParams}
\alias{GetJpegOrPngX11Params}
\alias{GetlimmaDataSetName}
\alias{GetMenuName}
\alias{GetMultipleContrasts}
\alias{GetNormalizationMethod}
\alias{getPackageVersion}
\alias{GetParameterizationName}
\alias{GetPlotLabels}
\alias{GetPlotSize}
\alias{GetPlotTitle}
\alias{GetRNATypesFrom.ContrastsFromDropDowns.String}
\alias{GetSetNames}
\alias{GetSlideNum}
\alias{GetSlideNums}
\alias{GetWhichProbes}
\alias{GetWtAreaParams}
\alias{GetWtFlagParams}
\alias{HeatDiagramDialog}
\alias{HeatDiagramPlot}
\alias{HTMLplotUsingFunction}
\alias{ImageArrayPlot}
\alias{ImageQualityPlot}
\alias{initGlobals}
\alias{IntensityHistogram}
\alias{IntensityHistogramAll}
\alias{LogOddsPlot}
\alias{NewLimmaFile}
\alias{NormalizedIntensityBoxPlot}
\alias{NormalizeNow}
\alias{nstrstr}
\alias{onExit}
\alias{OpenCDFandTargetsfiles}
\alias{OpenCDFFile}
\alias{OpenALimmaFile}
\alias{OpenLimmaFile}
\alias{OpenTargetsFile}
\alias{QQTplot}
\alias{RawIntensityBoxPlot}
\alias{Require}
\alias{Resize}
\alias{SaveAsLimmaFile}
\alias{SaveGraphAsJpeg}
\alias{SaveGraphAsPDF}
\alias{SaveGraphAsPNG}
\alias{SaveGraphAsPostscript}
\alias{SaveLimmaFile}
\alias{SetupPlotKeyBindings}
\alias{SetupPlotMenus}
\alias{SetWD}
\alias{showTopTable}
\alias{SimplifyContrastsExpression}
\alias{strstr}
\alias{tclArrayVar}
\alias{TclRequire}
\alias{toptable2}
\alias{topTable2}
\alias{Try}
\alias{TryReadImgProcFile}
\alias{UpDownOrBoth}
\alias{vennDiagramaffylmGUI}
\alias{VennDiagramPlot}
\alias{ViewContrastsMatrixAsPairs}
\alias{ViewContrastsMatrixInTable}
\alias{ViewExistingContrastParameterization}
\alias{ViewRNATargets}
\alias{affyHelp}
\alias{affylmGUIhelp}
\alias{limmaHelp}
\alias{ALGlimmaUsersGuide}
\alias{showCitations}
\alias{showChangeLog}
\alias{onDestroy}
\alias{GetPValueCutoff}
\alias{fixSeps}
\alias{PlotOptions}
\description{Graphical User Interface for Affy analysis using the limma microarray package}}

\usage{
AboutaffylmGUI()
AboutCustomMenu()
AboutNormalization()
addMenuItem()
affylmGUI()
affyPlotMA()
affyPlotMAcontrast()
ChooseCDF()
ChooseContrastParameterization()
chooseDir()
ChooseEbayesStatistic()
ComputeContrasts()
ComputeLinearModelFit()
CopyGraph()
DeleteContrastParameterization()
DeleteCustomMenuItem()
deleteItemFromList()
DensityPlot()
DensityPlotAll()
RNADigestionPlotAll()
NUSEPlotAll()
RLEPlotAll()
evalRcode()
ExportTopTable()
ExportHTMLreport()
ExportNormalizedExpressionValues()
generalPlotFunction()
GetComponentsToExportInHTMLreport()
GetContrast()
GetContrastNamesForHeatDiagram()
GetContrastParameterizationName()
GetContrasts()
GetCustomMenuItemNum()
GetDEcutoff()
GetGeneLabelsOptions()
GetJpegOrPngParams()
GetJpegOrPngX11Params()
GetlimmaDataSetName()
GetMenuName()
GetMultipleContrasts()
GetNormalizationMethod()
getPackageVersion()
GetParameterizationName()
GetPlotLabels()
GetPlotSize()
GetPlotTitle()
GetRNATypesFrom.ContrastsFromDropDowns.String()
GetSetNames()
GetSlideNum()
GetSlideNums()
GetWhichProbes()
GetWtAreaParams()
GetWtFlagParams()
HeatDiagramDialog()
HeatDiagramPlot()
HTMLplotUsingFunction()
ImageArrayPlot()
ImageQualityPlot()
initGlobals()
IntensityHistogram()
IntensityHistogramAll()
LogOddsPlot()
NewLimmaFile()
NormalizedIntensityBoxPlot()
NormalizeNow()
nstrstr()
onExit()
OpenCDFandTargetsfiles()
OpenCDFFile()
OpenALimmaFile()
OpenLimmaFile(FileName)
OpenTargetsFile()
QQTplot()
RawIntensityBoxPlot()
Require()
Resize()
SaveAsLimmaFile()
SaveGraphAsJpeg()
SaveGraphAsPDF()
SaveGraphAsPNG()
SaveGraphAsPostscript()
SaveLimmaFile()
SetupPlotKeyBindings()
SetupPlotMenus()
SetWD()
showTopTable()
SimplifyContrastsExpression()
strstr()
tclArrayVar()
TclRequire()
toptable2()
topTable2()
Try()
TryReadImgProcFile()
UpDownOrBoth()
vennDiagramaffylmGUI()
VennDiagramPlot()
ViewContrastsMatrixAsPairs()
ViewContrastsMatrixInTable()
ViewExistingContrastParameterization()
ViewRNATargets()
fixSeps(string)
affylmGUI(BigfontsForaffylmGUIpresentation=FALSE)
affyHelp()
affylmGUIhelp()
limmaHelp()
ALGlimmaUsersGuide()
showCitations()
showChangeLog()
onDestroy()
GetPValueCutoff(p.value=0.01)
PlotOptions()
}


\arguments{
	\item{BigfontsForaffylmGUIpresentation}{If set to TRUE, larger fonts are used.  However, some font
	sizes are not controlled by affylmGUI and so must be adjusted in the operating system, e.g.
	in the Control Panel in Windows under Display, Appearance.}
	\item{p.value}{A p-value cutoff.}
	\item{FileName}{A file name.}
	\item{string}{A character string.}
}

\details{
This function launches a Graphical User Interface for the limma package by Gordon Smyth.
The GUI uses Tk widgets (via the R TclTk interface by Peter Dalgaard) in order to provide
a simple interface to the limma functions for linear modelling of microarrays and
identification of differentially expressed genes.
}

\author{James Wettenhall}

\keyword{IO}
